chore: rename new-guardians-list to new-address-list
